backend/pyproject.toml
[project]
name = "xautopost-backend"
version = "0.0.1"
description = "Xautopost Python sidecar (FastAPI + Playwright)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "cryptography>=43.0.0",
    "argon2-cffi>=23.1.0",
    "keyring>=25.0.0",
    "playwright>=1.48.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
    "openai>=1.55.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pyinstaller>=6.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
